Add `AberthSolver` and allow complex coefficients
add missing cfg

remove needless traits

update readme & fix clippy warnings

return iteration count in `Roots`

make `ComplexCoefficient` & `sample_polynomial` private

unify return values from `aberth` and `AberthSolver::find_roots`

add documentation for `AberthSolver`, `Roots`, `StopReason`

add `to_owned` method to `Roots` containing references

re-export `num_complex::Complex`

move all items for `std` feature into a `feature_std` module

move tests into their own file module

add safety comments

return initial guesses when max_iterations is zero
